requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.0
numpy==1.26.2
scikit-learn==1.3.2
django-cors-headers==4.3.0
drf-yasg==1.21.7
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-filter==23.3
pydub==0.25.1
gTTS==2.3.2

